[Feat] 이미지 가져오기 추가

+++ src/main/java/kr/inuappcenterportal/inuportal/service/RedisService.java
@@ -59,7 +59,7 @@ public class RedisService {
         }
     }
 
-    public byte[] findImages(Long postId, Integer imageId){
+    public byte[] findImages(Long postId, Long imageId){
         String key = postId + "-" + imageId;
         log.info("이미지가져오기 key:{}",key);
         byte[] image = redisTemplateForImage.opsForValue().get(key);
